+++ Source/BlueprintExtensionsRuntime/Public/Features/SupportSystemLibrary.h
@@ -2,6 +2,7 @@
 
 #include "CoreMinimal.h"
 #include "Kismet/BlueprintFunctionLibrary.h"
+#include "Engine/EngineTypes.h"
 #include "SupportSystemLibrary.generated.h"
 
 UCLASS()